Separate channel.js
